+++ src/parser/item/index.js
@@ -179,10 +179,6 @@ function enrichFlags(ddbItem, item) {
   if (ddbItem.definition?.stackable) item.flags.ddbimporter.dndbeyond['stackable'] = ddbItem.definition.stackable;
 }
 
-function checkIfStringStartsWithLowerCase(str, subStrings) {
-  return subStrings.some((subString) => str.toLowerCase().startsWith(subString.toLowerCase()));
-}
-
 // the filter type "Other Gear" represents the equipment while the other filters represents the magic items in ddb
 export function parseItem(ddb, ddbItem, character, flags) {
   try {